stm32l4: move stm32l4 internal functions and definition into stm inside
STM32L4 specific datas are better to locate STM32L4 folder inside.
struct stm32l4_dsi_s, struct stm32l4_lcd_s, struct stm32l4_lcd_s *stm32l4_lcd_function(void);
