+++ os/arch/arm/src/stm32l4/stm32l4_dsi.h
@@ -79,14 +79,11 @@
 /************************************************************************************
  * Public Types
  ************************************************************************************/
-#if 0
-struct dsi_ops_s {
-    uint8_t (*config)(void);
-    void (*enable)(void);
-    void (*disable)(void);
-    void (*refresh)(void);
+struct stm32l4_dsi_s {
+    struct dsi_ops_s g_dsiops;
+    uint32_t isframebuffer;
 };
-#endif
+
 /************************************************************************************
  * Public Data
  ************************************************************************************/
