+++ os/board/stm32l4r9ai-disco/src/stm32_lcd.c
@@ -25,16 +25,15 @@
 #include <tinyara/arch.h>
 #include <tinyara/board.h>
 #include <tinyara/spi/spi.h>
-#include <tinyara/lcd/ieg1120.h>
-#include <tinyara/lcd/lcd.h>
-
-#include "chip/stm32l4_dma2d.h"
 
 #include <arch/board/board.h>
 #include <arch/board/stm32l4r9ai-disco.h>
 #include <arch/board/stm32l4r9i_discovery.h>
 
 #include <stm32l4.h>
+
+#include "chip/stm32l4_dma2d.h"
+#include "stm32l4_dsi.h"
 
 #define BRIGHTNESS_MIN      50
 #define BRIGHTNESS_NORMAL   200
