+++ os/board/stm32l4r9ai-disco/src/stm32l4r9ai-disco.h
@@ -201,6 +201,20 @@ extern struct spi_dev_s *g_spi1;
 extern struct spi_dev_s *g_spi2;
 #endif
 
+struct stm32l4_lcd_s {
+	/*  to control STM32L4R9I DISCOVERY Board Round LCD
+	 *
+	 *  - sendsparam    Send a multi length parameter for LCD configuration
+	 *  - sendlparam    Send a multi length parameter for LCD configuration
+	 */
+	uint8_t (*sendsparam)(uint32_t Mode, uint32_t Param1, uint32_t Param2);
+	uint8_t (*sendlparam)(uint32_t Mode, uint32_t NbParams, uint32_t Param1, uint8_t *ParametersTable);
+	void    (*enablewap)(void);
+	void    (*poweron)(void);
+	void    (*poweroff)(void);
+	void    (*refresh)(void);
+};
+
 /************************************************************************************
  * Public Functions
  ************************************************************************************/
@@ -274,5 +288,17 @@ void stm32l4_usbinitialize(void);
  *   Called to setup LTDC and DSI.
  *
  ************************************************************************************/
+
 void stm32l4_lcdinitialize(void);
+
+/************************************************************************************
+ * Name: stm32l4_lcd_function
+ *
+ * Description:
+ *   Called to set LCD Lower-half drivers
+ *
+ ************************************************************************************/
+
+struct stm32l4_lcd_s *stm32l4_lcd_function(void);
+
 #endif /* __CONFIGS_STM32L4R9AI_DISCO_SRC_STM32L4R9AI_DISCO_H */
