+++ os/board/stm32l4r9ai-disco/src/stm32l4r9i_lcd.c
@@ -44,11 +44,11 @@
 #include <tinyara/arch.h>
 #include <tinyara/board.h>
 #include <arch/board/board.h>
-#include <tinyara/lcd/ieg1120.h>
-#include <tinyara/lcd/lcd.h>
 
 #include <arch/board/stm32l4r9ai-disco.h>
 #include <arch/board/stm32l4r9i_discovery_io.h>
+
+#include "stm32l4_dsi.h"
 
 /****************************************************************************
  * Private Types
